project/requirements.txt
alembic==1.7.7
asyncpg==0.25.0
bcrypt==3.2.0
fastapi==0.75.1
python-jose==3.3.0
pydantic==1.9.0
passlib==1.7.4
sqlalchemy==1.4.35
uvicorn==0.17.6
loguru==0.6.0